Change from usize to u64
On [random-access-storage](https://github.com/datrs/random-access-storage)
issue
https://github.com/datrs/random-access-storage/issues/6 has changed all
types from `usize` into `u64` to be able to handle more than 4gbs on
32bits systems.

> usize is 32 bits on a 32 bit system, so the storage would be limited
to 4GB on such a system.

When changing `random-access-storage` on `hypercore` (tracking:
https://github.com/datrs/hypercore/pull/100) one of the things I've
noticed is that it wold be easier to also make `tree-index` use
`u64` to integrate with `random-access-storage`. Very likely, this also
means that we would need to bump to use more than 32Gb storages on
`tree-index`.

I've simply changed all declarations of usize into u64. All tests are
passing.

It also fixes Travis CI and update libraries to latest version.

Needs:
---

- [ ] https://github.com/datrs/flat-tree/pull/44

+++ src/proof.rs
@@ -3,9 +3,9 @@
 /// Merkle trees are proven by checking the parent hashes.
 #[derive(Debug, PartialEq)]
 pub struct Proof<'a> {
-  index: usize,
-  verified_by: usize,
-  nodes: &'a [usize],
+  index: u64,
+  verified_by: u64,
+  nodes: &'a [u64],
 }
 
 impl<'a> Proof<'a> {
@@ -19,7 +19,7 @@ impl<'a> Proof<'a> {
   /// let _proof = Proof::new(0, 0, &nodes);
   /// ```
   #[inline]
-  pub fn new(index: usize, verified_by: usize, nodes: &'a [usize]) -> Self {
+  pub fn new(index: u64, verified_by: u64, nodes: &'a [u64]) -> Self {
     Self {
       index,
       nodes,
@@ -38,7 +38,7 @@ impl<'a> Proof<'a> {
   /// assert_eq!(proof.index(), 0);
   /// ```
   #[inline]
-  pub fn index(&self) -> usize {
+  pub fn index(&self) -> u64 {
     self.index
   }
 
@@ -53,7 +53,7 @@ impl<'a> Proof<'a> {
   /// assert_eq!(proof.verified_by(), 0);
   /// ```
   #[inline]
-  pub fn verified_by(&self) -> usize {
+  pub fn verified_by(&self) -> u64 {
     self.verified_by
   }
 
@@ -68,7 +68,7 @@ impl<'a> Proof<'a> {
   /// assert_eq!(proof.nodes().len(), 0);
   /// ```
   #[inline]
-  pub fn nodes(&self) -> &[usize] {
+  pub fn nodes(&self) -> &[u64] {
     &self.nodes
   }
 }
